// utils/types.ts
export interface metaTags {
  pageTitle: string
  description: string
  siteName?: string
  contentType: string
  url: string
  socialBanner?: string
  bannerWidth?: string
  bannerHeight?: string
  twitterSiteName?: string
  twitterCardType?: string
  twitterUrl: string
  twitterTitle: string
  twitterDescription: string
  twitterSocialBanner?: string
}